+++ Package.swift
@@ -19,11 +19,11 @@ let package = Package(
         .executable(
             name: "GenerateEmbeddedDTDs",
             targets: ["GenerateEmbeddedDTDs"]),
-    ],
+   ],
     dependencies: [
         .package(url: "https://github.com/apple/swift-argument-parser", from: "1.6.0"),
         .package(url: "https://github.com/orchetect/swift-timecode", from: "3.0.0"),
-        .package(url: "https://github.com/orchetect/swift-extensions", from: "2.0.0"),
+        .package(url: "https://github.com/orchetect/swift-extensions", from: "2.1.0"),
         .package(url: "https://github.com/apple/swift-log.git", from: "1.0.0"),
     ],
     targets: [